fix the "k" value type; int -> double

+++ src/main/java/tokyo/nakanaka/roseCurveParticle/commandHandler/CommandHandlerFunctions.java
@@ -55,7 +55,7 @@ public class CommandHandlerFunctions {
 		}
 		{//k
 			String strValue = "";
-			Integer value = task.getAngularVelocity();
+			Double value = task.getAngularVelocity();
 			if(value != null) {
 				strValue = value.toString();
 			}
